Create ShapesNonMinified.js
This is the non minified version of my code, this will not work in a bookmarklet as when it is condensed (as it is here) into one line it will throw an error. But this version will work if ran as is. There was no real reason to include this but I just wanted to because this is my first "game".